backend/requirements.txt
# InventorySync Backend Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database & ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Async support
asyncpg==0.29.0
aiosqlite==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyjwt==2.8.0
cryptography==41.0.7
bcrypt==4.1.2

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Clients for Shopify API
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Background Tasks & Queue
celery[redis]==5.3.4
redis[hiredis]==5.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Forecasting & Analytics
scikit-learn==1.3.2
statsmodels==0.14.1

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.40.0
psutil==5.9.6
python-json-logger==2.0.7

# Caching
hiredis==2.3.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
safety==3.0.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production
prometheus-client==0.19.0
grpcio==1.60.0
